move dialog helper to popup_utils
