+++ browser/popup_handler_utility.py
@@ -4,8 +4,8 @@ import utils
 from .popup_handler import (
     setup_dialog_handler as _setup_dialog_handler,
     register_dialog_handler,
-    add_safe_accept_once,
 )
+from . import popup_utils
 
 from popup_text_handler import handle_popup_by_text
 
@@ -43,7 +43,7 @@ def close_all_popups_event(page: Page, loops: int = 2, wait_ms: int = 1000) -> b
                 if not btn.is_visible():
                     continue
                 try:
-                    add_safe_accept_once(page)
+                    popup_utils.add_safe_accept_once(page)
                     with page.expect_popup(timeout=500) as pop:
                         btn.click(timeout=0)
                     if pop.value:
@@ -141,7 +141,7 @@ def close_layer_popup(
             try:
                 btn = page.locator(sel)
                 if btn.count() > 0 and btn.first.is_visible():
-                    add_safe_accept_once(page)
+                    popup_utils.add_safe_accept_once(page)
                     with page.expect_popup(timeout=500) as pop_info:
                         btn.first.click()
                     if pop_info.value:
